# openapi.yaml
openapi: 3.0.1
info:
  title: API Gestion Mairie
  description: 'API REST pour la gestion des réservations de biens municipaux, des
    utilisateurs, des groupes et des pièces justificatives. Système de coordination
    pour la mairie.'
  contact:
    name: Équipe Mairie
    email: contact@mairie.fr
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Serveur de développement local
security:
- bearer-jwt: []
...
paths:
...
  /groupes:
    get:
      tags:
      - Groupes
      summary: Liste tous les groupes
      description: Récupère la liste complète des groupes (ADMIN/MODERATEUR)
      operationId: getAllGroupes
      responses:
        '200':
          description: Liste des groupes récupérée
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Groupe'
        '403':
          description: Accès refusé
      security:
      - bearer-jwt: []
    post:
      tags:
      - Groupes
      summary: Crée un nouveau groupe
      description: Crée un groupe d'utilisateurs avec permissions déléguées
      operationId: createGroupe
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Groupe'
        required: true
      responses:
        '201':
          description: Groupe créé avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Groupe'
        '403':
          description: Accès refusé
      security:
      - bearer-jwt: []
  /groupes/{id}:
    get:
      tags:
      - Groupes
      summary: Récupère un groupe par ID
      description: Retourne les détails d'un groupe spécifique (ADMIN/MODERATEUR)
      operationId: getGroupeById
      parameters:
      - name: id
        in: path
        description: ID du groupe
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Groupe trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Groupe'
        '403':
          description: Accès refusé
        '404':
          description: Groupe non trouvé
      security:
      - bearer-jwt: []
    delete:
      tags:
      - Groupes
      summary: Supprime un groupe
      description: Supprime un groupe (sauf CONSEIL_MUNICIPAL) avec permissions déléguées
      operationId: deleteGroupe
      parameters:
      - name: id
        in: path
        description: ID du groupe
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Groupe supprimé avec succès
        '403':
          description: Accès refusé
        '404':
          description: Groupe non trouvé
      security:
      - bearer-jwt: []
  /reservations/{id}/statut:
    patch:
      tags:
      - Réservations
      summary: Valide une réservation
      description: Change le statut de validation d'une réservation avec permissions
        déléguées
      operationId: updateReservationValidation
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateValidationRequest'
        required: true
      responses:
        '200':
          description: Statut mis à jour avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationDTO'
        '403':
          description: Accès refusé
        '404':
          description: Réservation non trouvée
      security:
      - bearer-jwt: []
  /reservations/{id}/caution:
    patch:
      tags:
      - Réservations
      summary: Modifie la caution d'une réservation
      description: Met à jour le statut de caution d'une réservation avec permissions
        déléguées
      operationId: updateReservationCaution
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCautionRequest'
        required: true
      responses:
        '200':
          description: Caution mise à jour avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationDTO'
        '403':
          description: Accès refusé
        '404':
          description: Réservation non trouvée
      security:
      - bearer-jwt: []
...
components:
  schemas:
...
    UserDTO:
      type: object
      properties:
        id_user:
          type: string
          format: uuid
        nom:
          type: string
        prenom:
          type: string
        email:
          type: string
        telephone:
          type: string
        role:
          type: string
          enum:
          - ADMIN
          - USER
          - MODERATEUR
          - TUTORED
          - SECRETARY
        statut:
          type: string
          enum:
          - ACTIF
          - INACTIF
          - SUSPENDU
          - PENDING
          - ACTIVE
          - REJECTED
    BienDTO:
      type: object
      properties:
        id_bien:
          type: string
          format: uuid
        nom:
          type: string
        description:
          type: string
        estVisible:
          type: boolean
        id_cat_bien:
          type: string
          format: uuid
        creerLe:
          type: string
          format: date-time
    Groupe:
      type: object
      properties:
        id_groupe:
          type: string
          format: uuid
        nom:
          type: string
        description:
          type: string
        type_groupe:
          type: string
          enum:
          - STANDARD
          - CONSEIL_MUNICIPAL
        type_exoneration:
          type: string
          enum:
          - ASSO
          - SOCIAL
          - MANDAT_ELECTIF
        niveau_tarif:
          type: integer
          format: int32
        creerLe:
          type: string
          format: date-time
    ReservationDTO:
      type: object
      properties:
        id_reservation:
          type: string
          format: uuid
        user:
          $ref: '#/components/schemas/UserDTO'
        bien:
          $ref: '#/components/schemas/BienDTO'
        date_debut:
          type: string
          format: date-time
        date_fin:
          type: string
          format: date-time
        est_valide:
          type: boolean
        statut_caution:
          type: string
          enum:
          - NON_REQUISE
          - EN_ATTENTE
          - VERSEE
          - RESTITUEE
    UpdateValidationRequest:
      type: object
      properties:
        est_valide:
          type: boolean
    UpdateCautionRequest:
      type: object
      properties:
        statut_caution:
          type: string
          enum:
          - NON_REQUISE
          - EN_ATTENTE
          - VERSEE
          - RESTITUEE
...
  securitySchemes:
    bearer-jwt:
      type: http
      description: 'Authentification JWT. Format: Bearer {token}'
      scheme: bearer
      bearerFormat: JWT